pyproject.toml
[project]
name = "dance-teacher"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "mcp>=1.13.1",
    "openai>=1.106.1",
    "langgraph>=0.2.50",
    "langchain-openai>=0.2.0",
    "langchain-core>=0.3.0",
    "python-dotenv>=1.0.0",
    "gradio>=4.0.0",
]
